# Fix condition simplifier