--- a/kore/src/Kore/Step/Simplification/Condition.hs
+++ b/kore/src/Kore/Step/Simplification/Condition.hs
@@ -15,38 +15,38 @@ module Kore.Step.Simplification.Condition (
 
 import Changed
 import qualified Control.Lens as Lens
-import Control.Monad.State.Strict (
-    StateT,
- )
+import Control.Monad.State.Strict
+    ( StateT
+    )
 import qualified Control.Monad.State.Strict as State
-import Data.Generics.Product (
-    field,
- )
+import Data.Generics.Product
+    ( field
+    )
 import qualified Kore.Internal.Condition as Condition
 import qualified Kore.Internal.Conditional as Conditional
 import qualified Kore.Internal.MultiAnd as MultiAnd
 import qualified Kore.Internal.OrPattern as OrPattern
-import Kore.Internal.Pattern (
-    Condition,
-    Conditional (..),
- )
+import Kore.Internal.Pattern
+    ( Condition
+    , Conditional (..)
+    )
 import qualified Kore.Internal.Pattern as Pattern
-import Kore.Internal.Predicate (
-    Predicate,
- )
+import Kore.Internal.Predicate
+    ( Predicate
+    )
 import qualified Kore.Internal.Predicate as Predicate
-import Kore.Internal.SideCondition (
-    SideCondition,
- )
+import Kore.Internal.SideCondition
+    ( SideCondition
+    )
 import qualified Kore.Internal.SideCondition as SideCondition
 import qualified Kore.Internal.Substitution as Substitution
-import Kore.Rewriting.RewritingVariable (
-    RewritingVariableName,
- )
+import Kore.Rewriting.RewritingVariable
+    ( RewritingVariableName
+    )
 import Kore.Step.Simplification.Simplify
-import Kore.Step.Simplification.SubstitutionSimplifier (
-    SubstitutionSimplifier (..),
- )
+import Kore.Step.Simplification.SubstitutionSimplifier
+    ( SubstitutionSimplifier (..)
+    )
 import qualified Kore.TopBottom as TopBottom
 import Kore.Unparser
 import Logic
@@ -84,7 +84,11 @@ simplify SubstitutionSimplifier{simplifySubstitution} sideCondition =
     worker Conditional{term, predicate, substitution} = do
         let substitution' = Substitution.toMap substitution
             predicate' = Predicate.substitute substitution' predicate
-        simplified <- simplifyPredicates sideCondition predicate'
+        partiallySimplified <- simplifyPredicate sideCondition predicate'
+        simplified <-
+            simplifyPredicates
+                sideCondition
+                (from @_ @(Predicate _) partiallySimplified)
         TopBottom.guardAgainstBottom simplified
         let merged = simplified <> Condition.fromSubstitution substitution
         normalized <- normalize merged