--- a/kore/src/Kore/Step/Simplification/Condition.hs
+++ b/kore/src/Kore/Step/Simplification/Condition.hs
@@ -84,11 +84,10 @@ simplify SubstitutionSimplifier{simplifySubstitution} sideCondition =
     worker Conditional{term, predicate, substitution} = do
         let substitution' = Substitution.toMap substitution
             predicate' = Predicate.substitute substitution' predicate
-        partiallySimplified <- simplifyPredicate sideCondition predicate'
+
         simplified <-
-            simplifyPredicates
-                sideCondition
-                (from @_ @(Predicate _) partiallySimplified)
+            simplifyPredicate sideCondition predicate'
+            >>= simplifyPredicates sideCondition . from @_ @(Predicate _)
         TopBottom.guardAgainstBottom simplified
         let merged = simplified <> Condition.fromSubstitution substitution
         normalized <- normalize merged