--- a/kore/test/Test/Kore/Step/Simplification/Pattern.hs
+++ b/kore/test/Test/Kore/Step/Simplification/Pattern.hs
@@ -111,7 +111,7 @@ test_Pattern_simplify =
                         (makeCeilPredicate fOfX)
                         ( makeExistsPredicate
                             Mock.xConfig
-                            (makeCeilPredicate (Mock.f (mkElemVar Mock.xConfig)))
+                            (makeCeilPredicate fOfX)
                         )
                     )
                     & OrPattern.fromPattern