@ana-pantilie 2021-04-08
@ttuegel Here is the data from the test runs, as we discussed in the meeting. I didn't generate the percentages, so I might have made mistakes, but the rest of the data should be accurate.

```
               Test name                              | Time PR | Time master |  Percentage  | Productivity on master
----------------------------------------------------------------------------------------------------------------------
tests/specs/functional/storageRoot-spec.k             |  1m  8s |    1m 25s   |  20% faster  |             - 
----------------------------------------------------------------------------------------------------------------------
tests/specs/functional/lemmas-spec.k                  |   55s   |    1m  3s   |  13% faster  |             - 
----------------------------------------------------------------------------------------------------------------------
tests/specs/examples/sum-to-n-spec.k                  |   2m    |    1m 40s   |  20% slower  |             - 
----------------------------------------------------------------------------------------------------------------------
tests/specs/functional/infinite-gas-spec.k            |   43s   |     47s     |   8% faster  |             - 
----------------------------------------------------------------------------------------------------------------------
tests/specs/functional/merkle-spec.k                  |   36s   |     38s     |   5% faster  |             - 
----------------------------------------------------------------------------------------------------------------------
tests/specs/functional/lemmas-no-smt-spec.k           |   39s   |     42s     |   7% faster  |             - 
----------------------------------------------------------------------------------------------------------------------
tests/specs/benchmarks/storagevar00-spec.k            |  7m 43s |    8m 10s   |   3% faster  |             - 
----------------------------------------------------------------------------------------------------------------------
tests/specs/benchmarks/overflow00-overflow-spec.k     | 11m 33s |   10m 11s   |  13% slower  |            90%
----------------------------------------------------------------------------------------------------------------------
tests/specs/benchmarks/storagevar03-spec.k            |  6m 36s |    6m  7s   |   8% slower  |            86%
----------------------------------------------------------------------------------------------------------------------
tests/specs/benchmarks/storagevar02-overflow-spec.k   | 11m  3s |   10m  9s   |   9% slower  |            86%
----------------------------------------------------------------------------------------------------------------------
tests/specs/benchmarks/storagevar02-nooverflow-spec.k | 11m 52s |   10m 54s   |   9% slower  |             -
----------------------------------------------------------------------------------------------------------------------
tests/specs/benchmarks/storagevar01-spec.k            | 10m 27s |    9m 28s   |  10% slower  |            84%
----------------------------------------------------------------------------------------------------------------------
tests/specs/erc20/hkg/totalSupply-spec.k              |  2m 46s |    2m 46s   |       -      |             -
----------------------------------------------------------------------------------------------------------------------
tests/specs/erc20/ds/totalSupply-spec.k               | 55m  5s |   59m 18s   |   7% faster  |             - 
```