@ttuegel 2021-04-08
I think that's fine. We see only small changes on the long-running tests. Those short tests have sudden speed-ups and slow-downs often, I'm not worried much about that. And anyway, we are working on the performance of all these tests right away.